+++ docs/versionhistory.rst
@@ -6,6 +6,7 @@ This library adheres to `Semantic Versioning 2.0 <https://semver.org/#semantic-v
 **UNRELEASED**
 
 - Fixed error message when a check against a ``Literal`` fails in a union on Python 3.10
+- Fixed ``NewType`` not being checked on Python 3.10
 
 **2.12.1** (2021-06-04)